refactor: consolidate redundant code
