+++ pyproject.toml
@@ -28,9 +28,7 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
 ]
 
-dependencies = [
-    "pyyaml>=6.0",
-]
+dependencies = []
 
 [project.optional-dependencies]
 triton = [
@@ -55,9 +53,6 @@ version_file = "src/flag_attn/_version.py"
 where = ["src"]
 include = ["flag_attn", "flag_attn.*"]
 namespaces = false
-
-[tool.setuptools.package-data]
-"flag_attn.runtime.backend._nvidia" = ["tune_configs.yaml"]
 
 # helps for setting up pytest in pyprojects
 # https://docs.pytest.org/en/7.3.x/reference/customize.html#rootdir
