+++ src/flag_attn/minimax_sparse_attention/index_topk.py
@@ -32,8 +32,7 @@ import triton
 import triton.language as tl
 from triton.errors import TritonError
 
-from ..utils import has_triton_tle
-from .utils import current_platform, round_up
+from flag_attn.utils import current_platform, has_triton_tle, round_up
 
 if has_triton_tle(3, 6, 0):
     try:
