+++ src/flag_attn/utils.py
@@ -123,3 +123,23 @@ def check_shared_mem(arch: str = "none", tensor_idx: int = 0) -> bool:
 def round_up(n: int, d: int) -> int:
     """Round n up to the nearest multiple of d."""
     return (n + d - 1) // d * d
+
+
+class _CurrentPlatform:
+    """Small platform adapter for backend capabilities shared by operators."""
+
+    def is_arch_support_pdl(self) -> bool:
+        """Return whether the active CUDA device supports launch PDL."""
+        if not torch.cuda.is_available():
+            return False
+        # HIP/ROCm does not accept CUDA's launch_pdl runtime argument.
+        if torch.version.hip is not None:
+            return False
+        try:
+            capability = torch.cuda.get_device_capability()
+        except RuntimeError:
+            return False
+        return capability >= (9, 0)
+
+
+current_platform = _CurrentPlatform()
